Remove scheme validation
URI.js's uri.protocol() interprets a raw hostname in the href attribute as a path element, and sets up "scheme:///path" later on for the href substitution. The uri.absoluteTo('https://') function does not handle this case well either, so remove the support for prefixing scheme to badly formed href values.